api: return user information when register successfully

+++ mandarin-common/src/main/java/org/a3/mandarin/common/util/PayUtil.java
@@ -3,7 +3,7 @@ package org.a3.mandarin.common.util;
 import org.a3.mandarin.common.exception.PayException;
 
 public class PayUtil {
-    public static void pay(Integer amount, String token) throws PayException{
+    public static void pay(Double amount, String token) throws PayException{
         // TODO: income mantain
     }
 }
